Delete /api/v1/posts/{id} 해당 포스트가 없는 경우, 작성한 사용자가 아닌 경우 예외 처리

+++ src/main/java/com/soonmin/final_project/service/PostService.java
@@ -10,6 +10,7 @@ import com.soonmin.final_project.exception.LikeLionException;
 import com.soonmin.final_project.repository.PostRepository;
 import com.soonmin.final_project.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.Authentication;
@@ -22,6 +23,7 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class PostService {
 
     private final PostRepository postRepository;
@@ -46,8 +48,18 @@ public class PostService {
         return postList.stream().map(post -> post.toDto().toViewResponse()).collect(Collectors.toList());
     }
 
-    public Integer delete(Integer id) {
+    public Integer delete(Integer id, Authentication authentication) {
+        Post post = postRepository.findById(id)
+                .orElseThrow(()->new LikeLionException(ErrorCode.POST_NOT_FOUND, ErrorCode.POST_NOT_FOUND.getMessage()));
+
+        log.info("userName get at post : {}", post.getUser().getUserName());
+        log.info("userName get at authentication : {}", authentication.getName());
+        if (!post.getUser().getUserName().equals(authentication.getName())){
+            throw new LikeLionException(ErrorCode.INVALID_PERMISSION, ErrorCode.INVALID_PERMISSION.getMessage());
+        }
+
         postRepository.deleteById(id);
+
         return id;
     }
 }
